database: Initial implementation of Inventory database
Inventory database supports the storage of records pertaining to the
hostgroups and the hosts associated with it. The database also supports
the backup to disk based storage once a fixed interval has been reached
so as to avoid any kind of data losses that may happen due to inadverant
service crashes.

Signed-off-by: Saurabh Badhwar <sbsaurabhbadhwar9@gmail.com>